Add graphing GUI
The plotting aspect is now embedded in its own class, and all three classes
are linked in multiple ways.

All three buttons are now hooked up to the Settings class, which now has a
button wrapper that will deactivate and activate the buttons when one of them
is clicked. They turn yellow to indicate that a calculation is happening.

Afterwards they all turn back, and the clicked button will turn red or green
depending on whether it succeeded or failed.